BE/User-server/requirements.txt
fastapi~=0.115.6
uvicorn~=0.34.0
PyMySQL~=1.1.1
SQLAlchemy~=2.0.37
python-dotenv~=1.0.1
pydantic~=2.10.5
bcrypt~=4.2.1
httpx~=0.28.1
